requirements.txt
boto3
PyPDF2
camelot-py
s3fs
flask